Relocate all code to /src
